LayoutTests/fast/canvas/imagedata-storageformat-enabled-expected.txt

Tests that ImageDataSettings contains storageFormat.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS created_imageData_uint8.width is 1
PASS created_imageData_uint8.height is 1
PASS created_imageData_uint8.data.constructor is Uint8ClampedArray
PASS created_imageData_uint8.data.BYTES_PER_ELEMENT is 1
PASS created_imageData_uint8.data.length is 4
PASS created_imageData_uint8.data.byteLength is 4
PASS created_imageData_uint8.data.at(0) is 0
PASS created_imageData_uint8.data.at(1) is 0
PASS created_imageData_uint8.data.at(2) is 0
PASS created_imageData_uint8.data.at(3) is 0
PASS gotten_imageData_uint8.width is 1
PASS gotten_imageData_uint8.height is 1
PASS gotten_imageData_uint8.data.constructor is Uint8ClampedArray
PASS gotten_imageData_uint8.data.BYTES_PER_ELEMENT is 1
PASS gotten_imageData_uint8.data.length is 4
PASS gotten_imageData_uint8.data.byteLength is 4
PASS gotten_imageData_uint8.data.at(0) is 0
PASS gotten_imageData_uint8.data.at(1) is 128
PASS gotten_imageData_uint8.data.at(2) is 255
PASS gotten_imageData_uint8.data.at(3) is 255
PASS created_imageData_float16.width is 1
PASS created_imageData_float16.height is 1
PASS created_imageData_float16.data.constructor is Float16Array
PASS created_imageData_float16.data.BYTES_PER_ELEMENT is 2
PASS created_imageData_float16.data.length is 4
PASS created_imageData_float16.data.byteLength is 8
PASS created_imageData_float16.data.at(0) is 0
PASS created_imageData_float16.data.at(1) is 0
PASS created_imageData_float16.data.at(2) is 0
PASS created_imageData_float16.data.at(3) is 0
FIXME: gotten_imageData_float16.data below should eventually become Float16Array.
PASS gotten_imageData_float16.width is 1
PASS gotten_imageData_float16.height is 1
PASS gotten_imageData_float16.data.constructor is Uint8ClampedArray
PASS gotten_imageData_float16.data.BYTES_PER_ELEMENT is 1
PASS gotten_imageData_float16.data.length is 4
PASS gotten_imageData_float16.data.byteLength is 4
PASS gotten_imageData_float16.data.at(0) is 0
PASS gotten_imageData_float16.data.at(1) is 128
PASS gotten_imageData_float16.data.at(2) is 255
PASS gotten_imageData_float16.data.at(3) is 255
PASS context.createImageData(1, 1, { storageFormat: "foo" }) threw exception TypeError: Type error.
PASS context.getImageData(0, 0, 1, 1, { storageFormat: "foo" }) threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
